fix reviver 2

+++ selection_sort/C_rand_work_lib/C_rand_work_lib_scripts/create_arrays.c
@@ -2,7 +2,8 @@
 #include <stdlib.h>
 #include "rand_main.h"
 
-float* create_rand_with_in_arr_f(int amount, float min,float max)
+
+float* create_rand_with_in_arr_f(size_t amount, float min,float max)
 {
     float* arr = malloc(sizeof(float)*amount);
     if(!arr)
@@ -15,7 +16,7 @@ float* create_rand_with_in_arr_f(int amount, float min,float max)
     }
     return arr;
 }
-double* rand_arr_create_with_in_d(int amount, float min,float max)
+double* rand_arr_create_with_in_d(size_t amount, float min,float max)
 {
     double* arr = malloc(sizeof(double)*amount);
     if(!arr)
@@ -28,7 +29,7 @@ double* rand_arr_create_with_in_d(int amount, float min,float max)
     }
     return arr;
 }
-int* rand_arr_create_with_in_int(int amount, int min,int max)
+int* rand_arr_create_with_in_int(size_t amount, int min,int max)
 {
     int* arr = malloc(sizeof(int)*amount);
     if(!arr)
@@ -41,7 +42,7 @@ int* rand_arr_create_with_in_int(int amount, int min,int max)
     }
     return arr;
 }
-char* rand_arr_create_with_in_char(int amount, char min,char max)
+char* rand_arr_create_with_in_char(size_t amount, char min,char max)
 {
     char* arr = malloc(sizeof(char)*(amount+1));
     if(!arr)
@@ -55,7 +56,7 @@ char* rand_arr_create_with_in_char(int amount, char min,char max)
     arr[amount]='\0';
     return arr;
 }
-char** rand_arr_create_with_in_string(int amount,int min_len,int max_len, char min,char max)
+char** rand_arr_create_with_in_string(size_t amount,int min_len,int max_len, char min,char max)
 {
     char** arr = malloc(sizeof(char*)*amount);
     if(!arr)
